% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{analyse_mixture_result}
\alias{analyse_mixture_result}
\title{Analyse mixture results}
\usage{
analyse_mixture_result(mix_res, unique_haps_in_mixture,
  unique_haps_in_mixture_counts)
}
\arguments{
\item{mix_res}{Mixture result from \code{\link[=mixture_info_by_individuals_2pers]{mixture_info_by_individuals_2pers()}},
\code{\link[=mixture_info_by_individuals_3pers]{mixture_info_by_individuals_3pers()}}, \code{\link[=mixture_info_by_individuals_4pers]{mixture_info_by_individuals_4pers()}},
\code{\link[=mixture_info_by_individuals_5pers]{mixture_info_by_individuals_5pers()}}}

\item{unique_haps_in_mixture}{Included unique haplotypes to use as elements in contributor tuples.}

\item{unique_haps_in_mixture_counts}{Population counts of the included haplotypes}
}
\value{
A list with numeric quantities
}
\description{
Calculate LR-like quantities by haplotype counts.
}
\details{
NOTE: Only takes up to 9 contributors!
}
