% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{analyse_mixture_results}
\alias{analyse_mixture_results}
\title{Analyse mixture results in a vectorised fashion}
\usage{
analyse_mixture_results(
  mixture_results,
  unique_haps_in_mixture_list,
  unique_haps_in_mixture_counts_list
)
}
\arguments{
\item{mixture_results}{List of \code{n} mixture results from \code{\link[=mixture_info_by_individuals_2pers]{mixture_info_by_individuals_2pers()}},
\code{\link[=mixture_info_by_individuals_3pers]{mixture_info_by_individuals_3pers()}}, \code{\link[=mixture_info_by_individuals_4pers]{mixture_info_by_individuals_4pers()}},
\code{\link[=mixture_info_by_individuals_5pers]{mixture_info_by_individuals_5pers()}}}

\item{unique_haps_in_mixture_list}{List of \code{n} included unique haplotypes, one for each element in \code{mix_res}}

\item{unique_haps_in_mixture_counts_list}{List of \code{n} population counts of the included unique haplotypes}
}
\value{
A list with lists of numeric quantities
}
\description{
Refer to \code{\link[=analyse_mixture_result]{analyse_mixture_result()}} for details.
Essentially, \code{\link[=analyse_mixture_result]{analyse_mixture_result()}} is run on each element of \code{mixture_results}.
}
\details{
NOTE: Only takes up to 9 contributors!
}
