% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{as_tbl_graph.malan_pedigreelist}
\alias{as_tbl_graph.malan_pedigreelist}
\title{Get tidy graph object}
\usage{
\method{as_tbl_graph}{malan_pedigreelist}(x, ...)
}
\arguments{
\item{x}{\code{malan_pedigreelist}}

\item{\dots}{Ignored}
}
\value{
\code{\link[=tbl_graph]{tbl_graph()}} object
}
\description{
Get tidy graph object \code{\link[=tbl_graph]{tbl_graph()}}, e.g.
to plot with \code{ggraph()}.
}
