% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{build_haplotype_hashmap}
\alias{build_haplotype_hashmap}
\title{Build hashmap of haplotype to individuals}
\usage{
build_haplotype_hashmap(individuals, progress = TRUE)
}
\arguments{
\item{individuals}{List of individuals to build hashmap of}

\item{progress}{Show progress?}
}
\value{
External pointer to hashmap with haplotype as keys and vector of individuals' pid as value
}
\description{
Makes it possible to find all individuals' pid with a certain haplotype.
Must be used with e.g. \code{\link[=get_matching_pids_from_hashmap]{get_matching_pids_from_hashmap()}}.
}
\seealso{
\code{\link[=get_matching_pids_from_hashmap]{get_matching_pids_from_hashmap()}}.
}
