% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{build_pedigrees}
\alias{build_pedigrees}
\title{Build pedigrees from (individuals in) a population.}
\usage{
build_pedigrees(population, progress = TRUE)
}
\arguments{
\item{population}{Population generated by \code{\link{sample_geneology}}.}

\item{progress}{Show progress.}
}
\value{
An object with class \code{malan_pedigreelist} (an internal list of external pointers to pedigrees).
}
\description{
In a newly simulated population, each individual only knows its father and children. 
Using this information, this function builds pedigrees.
}

