% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_autosomal_genotype_conditional_cumdist}
\alias{calc_autosomal_genotype_conditional_cumdist}
\title{Calculate conditional genotype cumulative probabilities with theta}
\usage{
calc_autosomal_genotype_conditional_cumdist(allele_dist, theta)
}
\arguments{
\item{allele_dist}{Allele distribution (probabilities) -- gets normalised}

\item{theta}{Theta correction between 0 and 1 (both included)}
}
\value{
Matrix: row i: conditional cumulative distribution of alleles given allele i
}
\description{
Calculate conditional genotype cumulative probabilities with theta
}
