% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_autosomal_genotype_probs}
\alias{calc_autosomal_genotype_probs}
\title{Calculate genotype probabilities with theta}
\usage{
calc_autosomal_genotype_probs(allele_dist, theta)
}
\arguments{
\item{allele_dist}{Allele distribution (probabilities) -- gets normalised}

\item{theta}{Theta correction between 0 and 1 (both included)}
}
\description{
Calculate genotype probabilities with theta
}
