% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship.R
\name{construct_M}
\alias{construct_M}
\title{Construct M matrix}
\usage{
construct_M(meioses, mu_dw, mu_up)
}
\arguments{
\item{meioses}{number of meioses separating the two individuals}

\item{mu_dw}{mutation rate for 1-step down-mutation}

\item{mu_up}{mutation rate for 1-step up-mutation}
}
\description{
Construct M matrix
}
