% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_haplotype_occurrences_individuals}
\alias{count_haplotype_occurrences_individuals}
\title{Count haplotypes occurrences in list of individuals}
\usage{
count_haplotype_occurrences_individuals(individuals, haplotype)
}
\arguments{
\item{individuals}{List of individuals to count occurrences in.}

\item{haplotype}{Haplotype to count occurrences of.}
}
\value{
Number of times that \code{haplotype} occurred amongst \code{individuals}.
}
\description{
Counts the number of types \code{haplotype} appears in \code{individuals}.
}
\seealso{
\code{\link[=pedigree_haplotype_matches_in_pedigree_meiosis_L1_dists]{pedigree_haplotype_matches_in_pedigree_meiosis_L1_dists()}}.
}
