% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_uncles}
\alias{count_uncles}
\title{Number of uncles}
\usage{
count_uncles(individual)
}
\arguments{
\item{individual}{individual}
}
\value{
Number of uncles
}
\description{
Get individual's number of uncles
}
