% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_autotheta_subpops_pids}
\alias{estimate_autotheta_subpops_pids}
\title{Estimate autosomal F, theta, and f from subpopulations of individual ids}
\usage{
estimate_autotheta_subpops_pids(population, subpops, subpops_sizes)
}
\arguments{
\item{population}{Population obtain from simulation}

\item{subpops}{List of individual pids}

\item{subpops_sizes}{Size of each subpopulation}
}
\value{
Estimates of autosomal F, theta, and f as well as additional information
}
\description{
Estimates autosomal F, theta, and f for a number of subpopulations given a list of pids (individual ids).
}
\details{
Assumes that \code{\link[=pedigrees_all_populate_autosomal]{pedigrees_all_populate_autosomal()}} was used first to populate autosomal genotypes.

Based on Bruce S Weir, Genetic Data Analysis 2, 1996. (GDA2).
}
