% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_autotheta_subpops_unweighted_genotypes}
\alias{estimate_autotheta_subpops_unweighted_genotypes}
\title{Unweighted estimate of autosomal theta from subpopulations of genotypes}
\usage{
estimate_autotheta_subpops_unweighted_genotypes(subpops, assume_HWE)
}
\arguments{
\item{subpops}{List of individual genotypes}

\item{assume_HWE}{if the alleles themselves are used instead of genotypes}
}
\value{
Estimate of autosomal theta
}
\description{
Estimates unweighted autosomal theta for a number of subpopulations given a list of subpopulations of genotypes.
}
\details{
Assumes that \code{\link[=pedigrees_all_populate_autosomal]{pedigrees_all_populate_autosomal()}} was used first to populate autosomal genotypes.

Based on Weir and Goudet, Genetics 2017:
http://www.genetics.org/content/early/2017/05/26/genetics.116.198424
}
