% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_theta_1subpop}
\alias{estimate_theta_1subpop}
\title{Estimate theta}
\usage{
estimate_theta_1subpop(x)
}
\arguments{
\item{x}{Matrix of genotypes: two columns (allele1 and allele2) and a row per individual}
}
\value{
Vector of length 1 containing estimate of theta or NA if it could not be estimated
}
\description{
Estimate theta for one subpopulation.
}
