% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_theta_1subpop_sample}
\alias{estimate_theta_1subpop_sample}
\title{Estimate theta from genetypes}
\usage{
estimate_theta_1subpop_sample(x)
}
\arguments{
\item{x}{Matrix of genotypes: two columns (allele1 and allele2) and a row per individual}
}
\value{
List:
\itemize{
\item estimate: Vector of length 1 containing estimate of theta or NA if it could not be estimated
\item error: true if an error happened, false otherwise
\item details: contains description if an error happened
}
}
\description{
Estimate theta for one subpopulation given a sample of genotypes.
}
