% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_theta_subpops_individuals}
\alias{estimate_theta_subpops_individuals}
\title{Estimate F, theta, and f from subpopulations of individuals}
\usage{
estimate_theta_subpops_individuals(subpops, subpops_sizes)
}
\arguments{
\item{subpops}{List of subpopulations, each a list of individuals}

\item{subpops_sizes}{Size of each subpopulation}
}
\value{
Estimates of F, theta, and f
}
\description{
Estimates F, theta, and f for a number of subpopulations given a list of individuals.
}
\details{
Based on Bruce S Weir, Genetic Data Analysis 2, 1996. (GDA2).
}
