% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_theta_subpops_pids}
\alias{estimate_theta_subpops_pids}
\title{Estimate F, theta, and f from subpopulations of individual ids}
\usage{
estimate_theta_subpops_pids(population, subpops, subpops_sizes)
}
\arguments{
\item{population}{Population obtain from simulation}

\item{subpops}{List of individual pids}

\item{subpops_sizes}{Size of each subpopulation}
}
\value{
Estimates of F, theta, and f
}
\description{
Estimates F, theta, and f for a number of subpopulations given a list of pids (individual ids).
}
\details{
Based on Bruce S Weir, Genetic Data Analysis 2, 1996. (GDA2).
}
