% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_theta_subpops_unweighted_genotypes}
\alias{estimate_theta_subpops_unweighted_genotypes}
\title{Unweighted estimate of theta from subpopulations of genotypes}
\usage{
estimate_theta_subpops_unweighted_genotypes(subpops, assume_HWE)
}
\arguments{
\item{subpops}{List of individual genotypes}

\item{assume_HWE}{if the alleles themselves are used instead of genotypes}
}
\value{
Estimate of theta
}
\description{
Estimates unweighted theta for a number of subpopulations given a list of subpopulations of genotypes.
}
\details{
Based on Weir and Goudet, Genetics 2017:
http://www.genetics.org/content/early/2017/05/26/genetics.116.198424
}
