% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_theta_subpops_unweighted_pids}
\alias{estimate_theta_subpops_unweighted_pids}
\title{Unweighted estimate of theta from subpopulations of individual ids}
\usage{
estimate_theta_subpops_unweighted_pids(population, subpops, assume_HWE)
}
\arguments{
\item{population}{Population obtain from simulation}

\item{subpops}{List of individual pids}

\item{assume_HWE}{if the alleles themselves are used instead of genotypes}
}
\value{
Estimate of theta
}
\description{
Estimates unweighted theta for a number of subpopulations given a list of pids (individual ids).
}
\details{
Based on Weir and Goudet, Genetics 2017:
http://www.genetics.org/content/early/2017/05/26/genetics.116.198424
}
