% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{father_matches}
\alias{father_matches}
\title{Father matches}
\usage{
father_matches(individual)
}
\arguments{
\item{individual}{individual}
}
\value{
Whether father has the same profile as \code{individual} or not
}
\description{
Does the father have the same profile as \code{individual}?
}
