% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{from_igraph_rcpp}
\alias{from_igraph_rcpp}
\title{Generate paternal brothers population}
\usage{
from_igraph_rcpp(vertices, edges)
}
\arguments{
\item{vertices}{vector of vertices}

\item{edges}{matrix with edges}
}
\value{
An external pointer to the population.
}
\description{
Generate paternal brothers population
}
