% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotypes.R
\name{generate_get_founder_haplotype}
\alias{generate_get_founder_haplotype}
\title{Generate a function to simulate pedigree founder haplotype}
\usage{
generate_get_founder_haplotype(ladder_min, ladder_max)
}
\arguments{
\item{ladder_min}{vector of minimum alleles; ladder_min[i] is the minimum allele at locus i}

\item{ladder_max}{vector of minimum alleles; ladder_max[i] is the maximum allele at locus i}
}
\description{
Generate a function to simulate pedigree founder haplotype
}
