% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotypes.R
\name{generate_get_founder_haplotype_db}
\alias{generate_get_founder_haplotype_db}
\title{Generate a function to simulate pedigree founder haplotype based on a haplotype databasep}
\usage{
generate_get_founder_haplotype_db(db)
}
\arguments{
\item{db}{data frame or matrix with haplotypes from which the founder is randomly simulated}
}
\description{
Generate a function to simulate pedigree founder haplotype based on a haplotype databasep
}
