% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_allele_counts_genotypes}
\alias{get_allele_counts_genotypes}
\title{Get autosomal allele counts from subpopulations of genotypes}
\usage{
get_allele_counts_genotypes(subpops)
}
\arguments{
\item{subpops}{List of individual genotypes}
}
\value{
Matrix with allele counts
}
\description{
Assumes that \code{\link[=pedigrees_all_populate_autosomal]{pedigrees_all_populate_autosomal()}} was used first to populate autosomal genotypes.
}
