% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_allele_counts_pids}
\alias{get_allele_counts_pids}
\title{Get allele counts from subpopulations given by pids}
\usage{
get_allele_counts_pids(population, subpops)
}
\arguments{
\item{population}{Population obtain from simulation}

\item{subpops}{List of individual pids}
}
\value{
Matrix with allele counts
}
\description{
Get allele counts from subpopulations given by pids
}
