% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_children}
\alias{get_children}
\title{Get children}
\usage{
get_children(individual)
}
\arguments{
\item{individual}{individual}
}
\value{
List with children
}
\description{
Get individual's children
}
\seealso{
\code{\link[=get_brothers]{get_brothers()}}, \code{\link[=get_uncles]{get_uncles()}}, \code{\link[=get_cousins]{get_cousins()}}
}
