% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_cousins}
\alias{get_cousins}
\title{Get cousins}
\usage{
get_cousins(individual)
}
\arguments{
\item{individual}{individual}
}
\value{
List with cousins
}
\description{
Get individual's cousins
}
\seealso{
\code{\link[=get_brothers]{get_brothers()}}, \code{\link[=get_uncles]{get_uncles()}}, \code{\link[=get_children]{get_children()}}
}
