% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_family_info}
\alias{get_family_info}
\title{Get individual's family information}
\usage{
get_family_info(individual)
}
\arguments{
\item{individual}{individual}
}
\value{
List with family information
}
\description{
Get individual's family information
}
