% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_father}
\alias{get_father}
\title{Get father}
\usage{
get_father(individual)
}
\arguments{
\item{individual}{individual}
}
\value{
Father
}
\description{
Get individual's father
}
\seealso{
\code{\link[=get_brothers]{get_brothers()}}, \code{\link[=get_uncles]{get_uncles()}}, \code{\link[=get_children]{get_children()}}, \code{\link[=get_cousins]{get_cousins()}}
}
