% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_generation}
\alias{get_generation}
\title{Get individual's generation number}
\usage{
get_generation(individual)
}
\arguments{
\item{individual}{Individual}
}
\value{
generation
}
\description{
Note that generation 0 is final, end generation.
1 is second last generation etc.
}
