% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_haplotypes_in_pedigree}
\alias{get_haplotypes_in_pedigree}
\title{Get haplotypes in pedigree}
\usage{
get_haplotypes_in_pedigree(ped)
}
\arguments{
\item{ped}{Pedigree}
}
\value{
List with haplotypes
}
\description{
Get haplotypes in pedigree
}
\examples{
sim <- sample_geneology(100, 10)
peds <- build_pedigrees(sim$population)
pedigrees_all_populate_haplotypes(peds, 2, c(1, 1))
get_haplotypes_in_pedigree(peds[[1]])

}
