% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_haplotypes_pids}
\alias{get_haplotypes_pids}
\title{Get haplotypes from a vector of pids.}
\usage{
get_haplotypes_pids(population, pids)
}
\arguments{
\item{population}{Population}

\item{pids}{Vector of pids to get haplotypes for.}
}
\value{
Matrix of haplotypes where row \code{i} is the haplotype of \code{individuals[[i]]}.

[[i]: R:[i
}
\description{
Requires that haplotypes are first populated, e.g.
with \code{\link[=pedigrees_all_populate_haplotypes]{pedigrees_all_populate_haplotypes()}},
\code{\link[=pedigrees_all_populate_haplotypes_custom_founders]{pedigrees_all_populate_haplotypes_custom_founders()}}, or
\code{\link[=pedigrees_all_populate_haplotypes_ladder_bounded]{pedigrees_all_populate_haplotypes_ladder_bounded()}}.
}
\seealso{
\code{\link[=get_haplotypes_individuals]{get_haplotypes_individuals()}}.
}
