% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_individuals}
\alias{get_individuals}
\title{Get all individuals in population}
\usage{
get_individuals(population)
}
\arguments{
\item{population}{Population}
}
\description{
Get all individuals in population
}
