% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_matching_pids_from_hashmap}
\alias{get_matching_pids_from_hashmap}
\title{Get individuals with a certain haplotype id by hashmap lookup}
\usage{
get_matching_pids_from_hashmap(hashmap, haplotype)
}
\arguments{
\item{hashmap}{Hashmap to make lookup in, made by \code{\link[=build_haplotype_hashmap]{build_haplotype_hashmap()}}}

\item{haplotype}{to get individuals that has this haplotype id}
}
\value{
List of individuals with a given haplotype id
}
\description{
By using hashmap made by \code{\link[=build_haplotype_hashmap]{build_haplotype_hashmap()}},
it is easy to get all individuals with a certain haplotype id.
}
\seealso{
\code{\link[=build_haplotype_hashmap]{build_haplotype_hashmap()}}.
}
