% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{get_nodes_edges}
\alias{get_nodes_edges}
\title{Get nodes and edges}
\usage{
get_nodes_edges(x, ...)
}
\arguments{
\item{x}{\code{malan_pedigreelist}}

\item{\dots}{Ignored}
}
\value{
List with entries \code{nodes} and \code{edges}
}
\description{
Get nodes and edges in \code{malan_pedigreelist}.
For example to plot via \code{\link[=as_tbl_graph]{as_tbl_graph()}}.
}
