% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_number_of_children}
\alias{get_number_of_children}
\title{Get number of children for each individual in the population}
\usage{
get_number_of_children(population, progress = TRUE)
}
\description{
Get number of children for each individual in the population
}

