% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_pedigree_as_graph}
\alias{get_pedigree_as_graph}
\title{Get pedigree information as graph (mainly intended for plotting)}
\usage{
get_pedigree_as_graph(ped)
}
\description{
Get pedigree information as graph (mainly intended for plotting)
}
