% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_pedigree_from_individual}
\alias{get_pedigree_from_individual}
\title{Get pedigree from individual}
\usage{
get_pedigree_from_individual(individual)
}
\arguments{
\item{individual}{Individual}
}
\value{
pedigree
}
\description{
Get pedigree from individual
}
