% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_pedigree_id}
\alias{get_pedigree_id}
\title{Get pedigree id}
\usage{
get_pedigree_id(ped)
}
\arguments{
\item{ped}{Pedigree}
}
\description{
Get pedigree id
}
\examples{
sim <- sample_geneology(100, 10)
peds <- build_pedigrees(sim$population)
get_pedigree_id(peds[[1]])

}
