% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_pedigrees_tidy}
\alias{get_pedigrees_tidy}
\title{Get pedigrees information in tidy format}
\usage{
get_pedigrees_tidy(pedigrees)
}
\arguments{
\item{pedigrees}{Pedigrees}
}
\description{
Get pedigrees information in tidy format
}
