% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_pid}
\alias{get_pid}
\title{Get pid from individual}
\usage{
get_pid(individual)
}
\arguments{
\item{individual}{Individual to get pid of}
}
\value{
pid
}
\description{
Get pid from individual
}
\examples{
sim <- sample_geneology(100, 10)
indv <- get_individual(sim$population, 1)
get_pid(indv)

}
