% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_pids_in_pedigree_criteria}
\alias{get_pids_in_pedigree_criteria}
\title{get pids in pedigree with certain criteria}
\usage{
get_pids_in_pedigree_criteria(ped, must_be_alive, use_birth_year,
  birth_year_min, birth_year_max)
}
\description{
get pids in pedigree with certain criteria
}

