% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_uncles}
\alias{get_uncles}
\title{Get uncles}
\usage{
get_uncles(individual)
}
\arguments{
\item{individual}{individual}
}
\value{
List with uncles
}
\description{
Get individual's uncles
}
\seealso{
\code{\link[=get_brothers]{get_brothers()}}, \code{\link[=get_children]{get_children()}}, \code{\link[=get_cousins]{get_cousins()}}
}
