% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotypes.R
\name{get_zero_haplotype_generator}
\alias{get_zero_haplotype_generator}
\title{Generate a function to generate the zero haplotype}
\usage{
get_zero_haplotype_generator(loci)
}
\arguments{
\item{loci}{Number of loci}
}
\description{
Generate a function to generate the zero haplotype
}
