% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grandfather_matches}
\alias{grandfather_matches}
\title{Grandfather matches}
\usage{
grandfather_matches(individual)
}
\arguments{
\item{individual}{individual}
}
\value{
Whether grandfather has the same profile as \code{individual} or not
}
\description{
Does the frandfather have the same profile as \code{individual}?
}
