% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{haplotype_matches_individuals}
\alias{haplotype_matches_individuals}
\title{Get individuals matching from list of individuals}
\usage{
haplotype_matches_individuals(individuals, haplotype)
}
\arguments{
\item{individuals}{List of individuals to count occurrences in.}

\item{haplotype}{Haplotype to count occurrences of.}
}
\value{
List of individuals that matches \code{haplotype} amongst \code{individuals}.
}
\description{
Get the indvididuals that matches \code{haplotype} in \code{individuals}.
}
\seealso{
\code{\link[=pedigree_haplotype_matches_in_pedigree_meiosis_L1_dists]{pedigree_haplotype_matches_in_pedigree_meiosis_L1_dists()}}.
}
