% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{haplotype_partially_matches_individuals}
\alias{haplotype_partially_matches_individuals}
\title{Get individuals partially matching from list of individuals}
\usage{
haplotype_partially_matches_individuals(
  individuals,
  haplotype,
  ignore_loci = as.integer(c())
)
}
\arguments{
\item{individuals}{List of individuals to count occurrences in.}

\item{haplotype}{Haplotype to count occurrences of.}

\item{ignore_loci}{Vector of loci to ignore (1 = ignore first locus etc.)}
}
\value{
List of individuals that partially matches \code{haplotype} amongst \code{individuals}.
}
\description{
Get the indvididuals that partially matches \code{haplotype} in \code{individuals}.
}
