% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{indices_in_mixture_by_haplotype_matrix}
\alias{indices_in_mixture_by_haplotype_matrix}
\title{Old, explicit IntegerMatrix haplotypes}
\usage{
indices_in_mixture_by_haplotype_matrix(haplotypes, H1, H2)
}
\description{
Old, explicit IntegerMatrix haplotypes
}
