% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{individual_pids_get_haplotypes}
\alias{individual_pids_get_haplotypes}
\title{Get haplotype matrix from list of individual pids}
\usage{
individual_pids_get_haplotypes(population, pids)
}
\description{
Get haplotype matrix from list of individual pids
}
