% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{individuals_get_haplotypes}
\alias{individuals_get_haplotypes}
\title{Get haplotype matrix from list of individuals}
\usage{
individuals_get_haplotypes(individuals)
}
\description{
Get haplotype matrix from list of individuals
}
