% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{infer_generation}
\alias{infer_generation}
\title{Infer individual's generation number}
\usage{
infer_generation(final_generation)
}
\arguments{
\item{final_generation}{Individuals in final generation}
}
\description{
Takes as input final generation, then moves up in pedigree and increments
generation number.
}
\details{
Note: Only works when all final generation individuals are provided.
}
