% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{infer_generations}
\alias{infer_generations}
\title{Infer generation numbers from pedigrees}
\usage{
infer_generations(peds)
}
\arguments{
\item{peds}{Pedigrees infered by \code{\link[=build_pedigrees]{build_pedigrees()}}}
}
\value{
Nothing
}
\description{
Infer generation numbers from pedigrees
}
