% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{load_haplotypes}
\alias{load_haplotypes}
\title{Load haplotypes to individuals}
\usage{
load_haplotypes(population, pid, haplotypes, progress = TRUE)
}
\arguments{
\item{pid}{ID of male}

\item{progress}{Show progress.}

\item{pid_dad}{ID of male's father, 0 if not known}

\item{error_on_pid_not_found}{Error if pid not found}
}
\description{
Note that individuals loaded this way does not have information about generation.
}
