% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{load_individuals}
\alias{load_individuals}
\title{Construct a population from data}
\usage{
load_individuals(pid, pid_dad, progress = TRUE, error_on_pid_not_found = TRUE)
}
\arguments{
\item{pid}{ID of male}

\item{pid_dad}{ID of male's father, 0 if not known}

\item{progress}{Show progress.}

\item{error_on_pid_not_found}{Error if pid not found}
}
\description{
Note that individuals loaded this way does not have information about generation.
}
