% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malan.R
\docType{package}
\name{malan-package}
\alias{malan}
\alias{malan-package}
\title{MAle Lineage ANlysis}
\description{
Simulating genealogies backwards and imposing STR mutations forwards.
}
\details{
Se vignettes and manual for documentation.

Disclaimer:
THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, TITLE AND NON-INFRINGEMENT.
IN NO EVENT SHALL THE COPYRIGHT HOLDERS OR ANYONE DISTRIBUTING
THE SOFTWARE BE LIABLE FOR ANY DAMAGES OR OTHER LIABILITY, WHETHER IN CONTRACT,
TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
}
\references{
Andersen MM, Balding DJ (2017) How convincing is a matching Y-chromosome profile?
PLoS Genet 13(11): e1007028. \url{https://doi.org/10.1371/journal.pgen.1007028}.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/mikldk/malan/issues}
}

}
\author{
\strong{Maintainer}: Mikkel Meyer Andersen \email{mikl@math.aau.dk}

}
