% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchprob.R
\name{matchprob}
\alias{matchprob}
\title{Calculate match probability information}
\usage{
matchprob(meiotic_dist, mut_rate)
}
\arguments{
\item{meiotic_dist}{Meiotic distribution (from \code{\link[=meiotic_distribution]{meiotic_distribution()}})}

\item{mut_rate}{Overall profile mutation rate}
}
\description{
Following \link{1},
$P(\text{match}) \approx \sum_{m = 1}^\infty
(1-\mu)^m \frac{P(M = m)}{P(M > 0)}$
for mutation probability $\mu$.

\link{1}: A. Caliebe and M. Krawczak (2018):
"Match probabilities for Y-chromosomal profiles: A paradigm shift"
}
