% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{meioses_generation_distribution}
\alias{meioses_generation_distribution}
\title{Meiotic distribution}
\usage{
meioses_generation_distribution(individual,
  generation_upper_bound_in_result = -1L)
}
\arguments{
\item{individual}{Individual to calculate all meiotic distances from}

\item{generation_upper_bound_in_result}{Limit on distribution; -1 means no limit.
0 is the final generation. 1 second last generation etc.}
}
\description{
Get the distribution of number of meioses from \code{individual}
to all individuals in \code{individual}'s pedigree.
Note the \code{generation_upper_bound_in_result} parameter.
}
