% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{meiotic_dist_threshold}
\alias{meiotic_dist_threshold}
\title{Meiotic distance between two individuals (with threshold)}
\usage{
meiotic_dist_threshold(ind1, ind2, threshold)
}
\arguments{
\item{ind1}{Individual 1}

\item{ind2}{Individual 2}

\item{threshold}{Max search radius, if exceeding, return -1}
}
\value{
Number of meioses between \code{ind1} and \code{ind2} if they are in the same pedigree, else -1.
}
\description{
Get the number of meioses between two individuals.
Note, that pedigrees must first have been inferred by \code{\link[=build_pedigrees]{build_pedigrees()}}.
}
