% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchprob.R
\name{meiotic_distribution}
\alias{meiotic_distribution}
\title{Infer meiotic distribution}
\usage{
meiotic_distribution(poi, indvs)
}
\arguments{
\item{poi}{Person of interest}

\item{indvs}{individuals in suspect population}
}
\description{
Infer meiotic distribution
}
