% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{meiotic_radius}
\alias{meiotic_radius}
\title{Meiotic radius}
\usage{
meiotic_radius(ind, radius)
}
\arguments{
\item{ind}{Individual}

\item{radius}{Max radius}
}
\value{
Matrix with ID and meiotic radius
}
\description{
Get all individual IDs within a meiotic radius
Note, that pedigrees must first have been inferred by \code{\link[=build_pedigrees]{build_pedigrees()}}.
}
