% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{meiotis_dist_all}
\alias{meiotis_dist_all}
\title{Find meiotic dist to all other individuals in pedigree}
\usage{
meiotis_dist_all(individual)
}
\arguments{
\item{individual}{Individual}
}
\description{
Find meiotic dist to all other individuals in pedigree
}
