% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{meiotis_dist_all_lookup}
\alias{meiotis_dist_all_lookup}
\title{Find meiotic dist to all other individuals in pedigree}
\usage{
meiotis_dist_all_lookup(individual)
}
\arguments{
\item{individual}{Individual}
}
\description{
Return an efficient look-up container
}
