% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixture_info_by_individuals}
\alias{mixture_info_by_individuals}
\title{New, implicit haplotypes by individuals}
\usage{
mixture_info_by_individuals(individuals, donor1, donor2)
}
\description{
New, implicit haplotypes by individuals
}
