% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixture_info_by_individuals_3pers}
\alias{mixture_info_by_individuals_3pers}
\title{Mixture information about 3 persons' mixture of donor1, donor2 and donor3.}
\usage{
mixture_info_by_individuals_3pers(individuals, donor1, donor2, donor3)
}
\arguments{
\item{individuals}{Individuals to consider as possible contributors and thereby get information from.}

\item{donor1}{Contributor1/donor 1}

\item{donor2}{Contributor2/donor 2}

\item{donor3}{Contributor2/donor 3}
}
\value{
A list with mixture information about the mixture \code{donor1}+\code{donor2}+\code{donor3} from \code{individuals}
}
\description{
Mixture information about 3 persons' mixture of donor1, donor2 and donor3.
}
\seealso{
\code{\link{mixture_info_by_individuals}}
}
