% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixture_info_by_individuals_5pers}
\alias{mixture_info_by_individuals_5pers}
\title{Mixture information about 5 persons' mixture of donor1, donor2, donor3, donor4 and donor5.}
\usage{
mixture_info_by_individuals_5pers(individuals, donor1, donor2, donor3,
  donor4, donor5)
}
\arguments{
\item{individuals}{Individuals to consider as possible contributors and thereby get information from.}

\item{donor1}{Contributor1/donor 1}

\item{donor2}{Contributor2/donor 2}

\item{donor3}{Contributor3/donor 3}

\item{donor4}{Contributor4/donor 4}

\item{donor5}{Contributor5/donor 5}
}
\value{
A list with mixture information about the mixture \code{donor1}+\code{donor2}+\code{donor3} from \code{individuals}
}
\description{
Mixture information about 5 persons' mixture of donor1, donor2, donor3, donor4 and donor5.
}
\seealso{
\code{\link{mixture_info_by_individuals_2pers}},
\code{\link{mixture_info_by_individuals_3pers}},
\code{\link{mixture_info_by_individuals_4pers}}
}
