% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{pedigree_as_igraph}
\alias{pedigree_as_igraph}
\title{Convert pedigree to igraph}
\usage{
pedigree_as_igraph(x, ...)
}
\arguments{
\item{x}{Pedigree}

\item{\dots}{ignored}
}
\value{
\code{igraph} object
}
\description{
Convert pedigree to igraph
}
