% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pedigree_size_generation}
\alias{pedigree_size_generation}
\title{Size of pedigree}
\usage{
pedigree_size_generation(pedigree, generation_upper_bound_in_result = -1L)
}
\arguments{
\item{pedigree}{Pedigree to get size of}

\item{generation_upper_bound_in_result}{Limit on generation to include in count; -1 means no limit.
0 only include the final generation. 1 only second last generation etc.}
}
\description{
Get the size of the pedigree.
Note the \code{generation_upper_bound_in_result} parameter.
}
