% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pedigrees_all_populate_autosomal_infinite_alleles}
\alias{pedigrees_all_populate_autosomal_infinite_alleles}
\title{Populate 1-locus autosomal DNA profile in pedigrees with infinite alleles mutation model.}
\usage{
pedigrees_all_populate_autosomal_infinite_alleles(population,
  mutation_rate, progress = TRUE)
}
\arguments{
\item{mutation_rate}{Mutation rate between 0 and 1 (both included)}

\item{progress}{Show progress}

\item{pedigrees}{Pedigree list in which to populate haplotypes}
}
\description{
Populate 1-locus autosomal DNA profile from founder and down in all pedigrees.
Note, that all founders have type 0 to begin with.
}
\details{
The maternal allele is taken by random from
the 2*N\link{g} alleles in the previous generation consisting of N\link{g} males
with descendants in the live population.

This is also why this is not using pedigrees but instead the population.

Note, that pedigrees need not be inferred.
}
\seealso{
\code{\link[=pedigrees_all_populate_haplotypes_custom_founders]{pedigrees_all_populate_haplotypes_custom_founders()}} and
\code{\link[=pedigrees_all_populate_haplotypes_ladder_bounded]{pedigrees_all_populate_haplotypes_ladder_bounded()}}.
}
