% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pedigrees_all_populate_haplotypes}
\alias{pedigrees_all_populate_haplotypes}
\title{Unbounded, founders all zero}
\usage{
pedigrees_all_populate_haplotypes(pedigrees, loci, mutation_rates,
  progress = TRUE)
}
\description{
Unbounded, founders all zero
}
