% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pedigrees_all_populate_haplotypes_custom_founders}
\alias{pedigrees_all_populate_haplotypes_custom_founders}
\title{Unbounded, custom founders}
\usage{
pedigrees_all_populate_haplotypes_custom_founders(pedigrees, mutation_rates,
  get_founder_haplotype = NULL, progress = TRUE)
}
\description{
Unbounded, custom founders
}
