% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pedigrees_all_populate_haplotypes_ladder_bounded}
\alias{pedigrees_all_populate_haplotypes_ladder_bounded}
\title{Bounded, custom founders}
\usage{
pedigrees_all_populate_haplotypes_ladder_bounded(pedigrees, mutation_rates,
  ladder_min, ladder_max, get_founder_haplotype = NULL, progress = TRUE)
}
\arguments{
\item{get_founder_haplotype}{has no default as it is not know in advance how many loci there are and what the ladder is; see \code{\link{generate_get_founder_haplotype}}}
}
\description{
Populate haplotypes such that they are all on-ladder
}
