% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pedigrees_count}
\alias{pedigrees_count}
\title{Get number of pedigrees}
\usage{
pedigrees_count(pedigrees)
}
\arguments{
\item{pedigrees}{Pedigrees}
}
\description{
Get number of pedigrees
}
\examples{
sim <- sample_geneology(100, 10)
peds <- build_pedigrees(sim$population)
pedigrees_count(peds)

}
