% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pedigrees_table}
\alias{pedigrees_table}
\title{Get distribution of pedigree sizes}
\usage{
pedigrees_table(pedigrees)
}
\arguments{
\item{pedigrees}{Pedigrees}
}
\description{
Get distribution of pedigree sizes
}
\examples{
sim <- sample_geneology(100, 10)
peds <- build_pedigrees(sim$population)
pedigrees_table(peds)

}
