% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{plot.malan_pedigree}
\alias{plot.malan_pedigree}
\title{Plot pedigree}
\usage{
\method{plot}{malan_pedigree}(x, ids = TRUE, haplotypes = FALSE,
  locus_sep = " ", mark_pids = NULL, label_color = "black",
  node_color = "lightgray", mark_color = "orange", ...)
}
\arguments{
\item{x}{Pedigree}

\item{ids}{Show pids}

\item{haplotypes}{Show haplotypes}

\item{locus_sep}{Locus separator in haplotypes}

\item{mark_pids}{Vector of pids to highlight}

\item{label_color}{Label color}

\item{node_color}{Node color}

\item{mark_color}{Highlight color}

\item{\dots}{Passed to \code{igraph::plot.igraph}}
}
\description{
Plot \code{malan_pedigree}.
}
