% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{plot.malan_pedigreelist}
\alias{plot.malan_pedigreelist}
\title{Plot pedigree list}
\usage{
\method{plot}{malan_pedigreelist}(x, ...)
}
\arguments{
\item{x}{Pedigree list}

\item{\dots}{ignored}
}
\description{
Plot \code{malan_pedigreelist} generated by
\code{\link[=build_pedigrees]{build_pedigrees()}}.
}
