% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{population_size_generation}
\alias{population_size_generation}
\title{Size of population}
\usage{
population_size_generation(population, generation_upper_bound_in_result = -1L)
}
\arguments{
\item{population}{Population to get size of}

\item{generation_upper_bound_in_result}{Limit on generation to include in count; -1 means no limit.
0 only include the final generation. 1 only second last generation etc.}
}
\description{
Get the size of the population.
Note the \code{generation_upper_bound_in_result} parameter.
}
