% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.malan_pedigree}
\alias{print.malan_pedigree}
\title{Print pedigree}
\usage{
\method{print}{malan_pedigree}(x, ...)
}
\arguments{
\item{x}{Pedigree}

\item{\dots}{ignored}
}
\description{
Print pedigree
}
