% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.malan_pedigreelist}
\alias{print.malan_pedigreelist}
\title{Print pedigree list}
\usage{
\method{print}{malan_pedigreelist}(x, ...)
}
\arguments{
\item{x}{Pedigrees (\code{malan_pedigreelist})}

\item{\dots}{ignored}
}
\description{
Print \code{malan_pedigreelist} generated by
\code{\link[=build_pedigrees]{build_pedigrees()}}.
}
