% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.malan_population}
\alias{print.malan_population}
\title{Print population}
\usage{
\method{print}{malan_population}(x, ...)
}
\arguments{
\item{x}{Population (\code{malan_population})}

\item{\dots}{ignored}
}
\description{
Print \code{malan_population} generated by
\code{\link[=sample_geneology]{sample_geneology()}} or \code{\link[=sample_geneology_varying_size]{sample_geneology_varying_size()}}.
}
