% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.malan_population_abort}
\alias{print.malan_population_abort}
\title{Print \code{malan_population_abort}}
\usage{
\method{print}{malan_population_abort}(x, ...)
}
\arguments{
\item{x}{\code{malan_population_abort}}

\item{\dots}{ignored}
}
\description{
Print \code{malan_population_abort}
}
