% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{print_individual}
\alias{print_individual}
\title{Print individual}
\usage{
print_individual(individual)
}
\arguments{
\item{individual}{Individual}
}
\description{
Print individual
}
\examples{
sim <- sample_geneology(100, 10)
indv <- get_individual(sim$population, 1)
print_individual(indv)

}
