% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship.R
\name{relationship_allele_diff_dist}
\alias{relationship_allele_diff_dist}
\title{Calculate distribution of allele difference}
\usage{
relationship_allele_diff_dist(meioses, mu_dw, mu_up)
}
\arguments{
\item{meioses}{number of meioses separating the two individuals}

\item{mu_dw}{mutation rate for 1-step down-mutation}

\item{mu_up}{mutation rate for 1-step up-mutation}
}
\value{
\code{data.frame} with columns \code{d} (allele difference) and \code{p} (prob)
}
\description{
Calculate distribution of allele difference after \code{m} meioses.
}
