% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_autosomal_genotype}
\alias{sample_autosomal_genotype}
\title{Sample genotype with theta}
\usage{
sample_autosomal_genotype(allele_dist, theta)
}
\arguments{
\item{allele_dist}{Allele distribution (probabilities) -- gets normalised}

\item{theta}{Theta correction between 0 and 1 (both included)}
}
\description{
Sample genotype with theta
}
