% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_generation}
\alias{set_generation}
\title{Set individual's generation number}
\usage{
set_generation(individual, generation)
}
\arguments{
\item{individual}{Individual}
}
\description{
Note that generation 0 is final, end generation.
1 is second last generation etc.
}
\examples{
sim <- sample_geneology(100, 10)
indv <- get_individual(sim$population, 1)
get_generation(indv)
set_generation(indv, 100)
get_generation(indv)

}
