% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{[[.malan_pedigreelist}
\alias{[[.malan_pedigreelist}
\title{Get pedigree from pedigree list}
\usage{
\method{[[}{malan_pedigreelist}(x, ...)
}
\arguments{
\item{x}{Element id}

\item{\dots}{ignored}
}
\value{
Pedigree
}
\description{
Get pedigree from \code{malan_pedigreelist} generated by
\code{\link[=build_pedigrees]{build_pedigrees()}}.
}
