% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{[[.malan_population}
\alias{[[.malan_population}
\title{Get individual from population by pid}
\usage{
\method{[[}{malan_population}(x, ...)
}
\arguments{
\item{x}{pid}

\item{\dots}{ignored}
}
\value{
Individual
}
\description{
Get individual from population by pid
}
