% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_create_population}
\alias{test_create_population}
\title{Generate test population}
\usage{
test_create_population()
}
\value{
An external pointer to the population.
}
\description{
Generate test population
}
