% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ystr_kits}
\alias{ystr_kits}
\title{Kit information about Y-STR markers}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{Marker}{name of Y-STR marker}
  \item{Kit}{name of Y-STR kit}
}}
\source{
\url{http://www.yhrd.org}
}
\usage{
ystr_kits
}
\description{
A dataset containing information about the 
Y chromosomal short tandem repeat (Y-STR) markers that are present in the kit.
}
\keyword{datasets}
