% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ystr_markers}
\alias{ystr_markers}
\title{Mutational information about Y-STR markers}
\format{A data frame with 53,940 rows and 10 variables:
\describe{
\item{Marker}{name of Y-STR marker}
\item{Meioses}{number of meioses observed}
\item{Mutations}{number of mutations observed in the corresponding number of Meioses}
\item{MutProb}{point estimate of mutation probability, MutProb = Mutations/Meioses}
\item{Alleles}{observed alleles}
}}
\source{
\url{http://www.yhrd.org}
}
\usage{
ystr_markers
}
\description{
A dataset from yhrd.org (and their sources) containing mutational information about
Y chromosomal short tandem repeat (Y-STR) markers used in forensic genetics.
}
\details{
Note, that loci with duplications (DYS385a/b as well as
DYF387S1a/b have been split into two loci).
}
\keyword{datasets}
