% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{brothers_matching}
\alias{brothers_matching}
\title{Number of brothers with matching haplotype}
\usage{
brothers_matching(individual)
}
\arguments{
\item{individual}{individual}
}
\value{
Number of brothers that matches \code{individual}'s haplotype
}
\description{
Get individual's number of brothers that matches \code{individual}'s haplotype
}
