% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_brothers}
\alias{count_brothers}
\title{Number of brothers}
\usage{
count_brothers(individual)
}
\arguments{
\item{individual}{individual}
}
\value{
Number of brothers
}
\description{
Get individual's number of brothers
}
\seealso{
\code{\link[=get_brothers]{get_brothers()}}
}
