% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_allele_counts_genotypes}
\alias{get_allele_counts_genotypes}
\title{Get allele counts from subpopulations of genotypes}
\usage{
get_allele_counts_genotypes(subpops)
}
\arguments{
\item{subpops}{List of individual genotypes}
}
\value{
Matrix with allele counts
}
\description{
Get allele counts from subpopulations of genotypes
}
