% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_haplotypes_in_pedigree}
\alias{get_haplotypes_in_pedigree}
\title{Get haplotypes in pedigree}
\usage{
get_haplotypes_in_pedigree(ped)
}
\arguments{
\item{ped}{Pedigree}
}
\value{
List with haplotypes
}
\description{
Get haplotypes in pedigree
}
