% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{load_haplotypes}
\alias{load_haplotypes}
\title{Load haplotypes to individuals}
\usage{
load_haplotypes(population, pid, haplotypes, progress = TRUE)
}
\arguments{
\item{population}{of individuals}

\item{pid}{ID of male}

\item{haplotypes}{\itemize{
\item row \code{i} has \code{pid[i]} ID
}}

\item{progress}{Show progress.}
}
\description{
Note that individuals loaded this way does not have information about generation.
}
