% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixture_info_by_individuals_3pers}
\alias{mixture_info_by_individuals_3pers}
\title{New, implicit haplotypes by individuals}
\usage{
mixture_info_by_individuals_3pers(individuals, donor1, donor2, donor3)
}
\description{
New, implicit haplotypes by individuals
}
