% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pedigrees_all_populate_haplotypes_ladder_bounded}
\alias{pedigrees_all_populate_haplotypes_ladder_bounded}
\title{Populate haplotypes in pedigrees (custom founder/bounded).}
\usage{
pedigrees_all_populate_haplotypes_ladder_bounded(pedigrees, mutation_rates,
  ladder_min, ladder_max, get_founder_haplotype = NULL, progress = TRUE)
}
\arguments{
\item{pedigrees}{Pedigree list in which to populate haplotypes}

\item{mutation_rates}{Vector with mutation rates}

\item{ladder_min}{Lower bounds for haplotypes, same length as \code{mutation_rates}}

\item{ladder_max}{Upper bounds for haplotypes, same length as \code{mutation_rates}; all entries must be strictly greater than \code{ladder_min}}

\item{get_founder_haplotype}{Function taking no arguments returning a haplotype of \code{length(mutation_rates)}}

\item{progress}{Show progress}
}
\description{
Populate haplotypes from founder and down in all pedigrees.
Note, that haplotypes are bounded by \code{ladder_min} and \code{ladder_max}.
All founders get a haplotype from calling the user
provided function \code{\link[=get_founder_haplotype]{get_founder_haplotype()}}.
}
\details{
Note, that pedigrees must first have been inferred by \code{\link[=build_pedigrees]{build_pedigrees()}}.
}
\seealso{
\code{\link[=pedigrees_all_populate_haplotypes]{pedigrees_all_populate_haplotypes()}} and
\code{\link[=pedigrees_all_populate_haplotypes_custom_founders]{pedigrees_all_populate_haplotypes_custom_founders()}}.
}
