% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship.R
\name{relationship_allele_diff_dist_sym}
\alias{relationship_allele_diff_dist_sym}
\title{Calculate distribution of allele difference for symmetric mutation rates}
\usage{
relationship_allele_diff_dist_sym(meioses, mu_updw)
}
\arguments{
\item{meioses}{number of meioses separating the two individuals}

\item{mu_updw}{mutation rate for 1-step down- and up-mutations, i.e. total mutation rate is \code{2*mu_updw}}
}
\value{
\code{data.frame} with columns \code{d} (allele difference) and \code{p} (prob)
}
\description{
Calculate distribution of allele difference after \code{m} meioses.
}
